+++ src/strigui/box.clj
@@ -66,3 +66,14 @@
               (when ticked?
                 (c/rect (+ x 5) (+ y 5) (- width 9) (- width 9) border-color true))
               (c/rect x y width width border-color false thickness))))
+
+(defn draw-circle
+  [context {:keys [x y width color thickness]} ticked?]
+  (let [background-color (get color :background (java.awt.Color. 250 250 250))
+        border-color (get color :border (java.awt.Color. 27 100 98))
+        [x2 y2] (mapv #(+ % (/ width 2)) [x y])]
+    (c/draw-> context
+              (c/ellipse x2 y2 width width background-color true)
+              (when ticked?
+                (c/ellipse x2 y2 (- width 9) (- width 9) border-color true))
+              (c/ellipse x2 y2 width width border-color false thickness))))