+++ src/strigui/playground.clj
@@ -48,7 +48,10 @@
   (default-stuff)
   (gui/swap-widgets! #(-> %
                           (gui/add-checkbox "main-window" "cb-none" nil {:x 10 :y 0 :text "Without a tick" :align-text :left})
-                          (gui/add-checkbox "main-window" "cb-tick" true {:x 10 :y 50 :text "With a tick" :align-text :left})))
+                          (gui/add-checkbox "main-window" "cb-tick" true {:x 10 :y 50 :text "With a tick" :align-text :left})
+                          (gui/add-radio "main-window" "cb-radio" nil {:x 10 :y 100 :text "Radio empty" :align-text :left})
+                          (gui/add-radio "main-window" "cb-radio-full" true {:x 10 :y 150 :text "Radio full" :align-text :left})))
+  
   ;;(gui/window! 200 300 1500 600 "My Window" (java.awt.Color. 44 44 44)) ;{java.awt.RenderingHints/KEY_ANTIALIASING java.awt.RenderingHints/VALUE_ANTIALIAS_ON}